[glib-demo] Add cast to make sure the comparison is always valid

+++ glib/demo/attachments.c
@@ -252,7 +252,7 @@ pgd_attachments_validate_button_clicked (GtkButton   *button,
 		gint i;
 	
 		for (i = 0; i < digest_len; i++) {
-			if (attachment->checksum->str[i] != digest[i]) {
+			if ((guint8)attachment->checksum->str[i] != digest[i]) {
 				valid = FALSE;
 				break;
 			}
